+++ tests/Service/CourseProviderTest.php
@@ -8,6 +8,8 @@ use ApiPlatform\Symfony\Bundle\Test\ApiTestCase;
 use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Apis\PublicRestCourseApi;
 use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\DependencyInjection\Configuration;
 use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\DependencyInjection\DbpRelayBaseCourseConnectorCampusonlineExtension;
+use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Entity\CachedCourse;
+use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Entity\CachedCourseTitle;
 use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\EventSubscriber\CourseEventSubscriber;
 use Dbp\Relay\BaseCourseConnectorCampusonlineBundle\Service\CourseProvider;
 use Dbp\Relay\CoreBundle\Exception\ApiError;
@@ -24,26 +26,31 @@ use Symfony\Component\EventDispatcher\EventDispatcher;
 class CourseProviderTest extends ApiTestCase
 {
     private const COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME = 'type';
-    private const COURSE_TYPE_SOURCE_ATTRIBUTE_NAME = 'type';
-    private const TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME = 'term';
-    private const TEACHING_TERM_SOURCE_ATTRIBUTE_NAME = 'teachingTerm';
+    private const COURSE_TYPE_LEGACY_SOURCE_ATTRIBUTE_NAME = 'type';
+    private const COURSE_TYPE_SOURCE_ATTRIBUTE_NAME = 'courseTypeKey';
+    private const SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME = 'term';
+    private const SEMESTER_LEGACY_SOURCE_ATTRIBUTE_NAME = 'teachingTerm';
+    private const SEMESTER_SOURCE_ATTRIBUTE_NAME = 'semesterKey';
 
     private ?CourseProvider $courseProvider = null;
     private ?EventDispatcher $eventDispatcher = null;
     private ?EntityManagerInterface $entityManager = null;
+    private ?CourseEventSubscriber $courseEventSubscriber = null;
 
-    private static function createEventSubscriberConfig(): array
+    private static function createEventSubscriberConfig(bool $publicRest): array
     {
         $config = [];
         $config['local_data_mapping'] = [
             [
                 'local_data_attribute' => self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME,
-                'source_attribute' => self::COURSE_TYPE_SOURCE_ATTRIBUTE_NAME,
+                'source_attribute' => $publicRest ?
+                    self::COURSE_TYPE_SOURCE_ATTRIBUTE_NAME : self::COURSE_TYPE_LEGACY_SOURCE_ATTRIBUTE_NAME,
                 'default_value' => '',
             ],
             [
-                'local_data_attribute' => self::TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME,
-                'source_attribute' => self::TEACHING_TERM_SOURCE_ATTRIBUTE_NAME,
+                'local_data_attribute' => self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME,
+                'source_attribute' => $publicRest ?
+                    self::SEMESTER_SOURCE_ATTRIBUTE_NAME : self::SEMESTER_LEGACY_SOURCE_ATTRIBUTE_NAME,
                 'default_value' => '',
             ],
         ];
@@ -61,38 +68,16 @@ class CourseProviderTest extends ApiTestCase
         $this->entityManager = TestEntityManager::setUpEntityManager($container,
             DbpRelayBaseCourseConnectorCampusonlineExtension::ENTITY_MANAGER_ID);
 
+        $this->createStagingTables();
+
         $this->courseProvider = new CourseProvider($this->entityManager, $this->eventDispatcher);
-        $this->courseProvider->setConfig($this->getPublicRestApiConfig());
+        // $this->courseProvider->setConfig($this->getPublicRestApiConfig());
         $this->courseProvider->setCache(new ArrayAdapter(), 3600);
         $this->courseProvider->setLogger(new NullLogger());
 
-        $localDataEventSubscriber = new CourseEventSubscriber($this->courseProvider);
-        $localDataEventSubscriber->setConfig(self::createEventSubscriberConfig());
-        $this->eventDispatcher->addSubscriber($localDataEventSubscriber);
-    }
-
-    private function getLegacyApiConfig(): array
-    {
-        $config = [];
-        $config[Configuration::CAMPUS_ONLINE_NODE] = [
-            'legacy' => true,
-            'org_root_id' => '1',
-        ];
-
-        return $config;
-    }
-
-    private function getPublicRestApiConfig(): array
-    {
-        $config = [];
-        $config[Configuration::CAMPUS_ONLINE_NODE] = [
-            'legacy' => false,
-            'base_url' => 'https://campusonline.at/campusonline/ws/public/rest/',
-            'client_id' => 'client',
-            'client_secret' => 'secret',
-        ];
-
-        return $config;
+        $this->courseEventSubscriber = new CourseEventSubscriber($this->courseProvider);
+        // $this->courseEventSubscriber->setConfig(self::createEventSubscriberConfig(true));
+        $this->eventDispatcher->addSubscriber($this->courseEventSubscriber);
     }
 
     private function mockResponses(array $responses, bool $mockAuthServerResponses = false): void
@@ -107,53 +92,192 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseByIdentifierEn(): void
     {
-        $coResponses = [
-            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
-                json_encode([
-                    'uid' => '240759',
-                    'title' => [
-                        'value' => [
-                            'en' => 'Computational Intelligence',
-                            'de' => 'Komputationswissenschaft',
-                        ],
-                    ],
-                ])),
-        ];
+        $this->setUpPublicRestApi();
 
         $options = [];
         Options::setLanguage($options, 'en');
-        $this->mockResponses($coResponses, true);
-        $course = $this->courseProvider->getCourseById('240759', $options);
-        $this->assertSame('240759', $course->getIdentifier());
-        $this->assertSame('Computational Intelligence', $course->getName());
+        $course = $this->courseProvider->getCourseById('2', $options);
+        $this->assertSame('2', $course->getIdentifier());
+        $this->assertSame('Computational Science', $course->getName());
     }
 
-    public function testGetCourseByIdentifierDe(): void
+    public function testGetCourseByIdentifierDeWithLocalData(): void
     {
+        $this->setUpPublicRestApi();
+
+        $options = [];
+        Options::setLanguage($options, 'de');
+        Options::requestLocalDataAttributes($options, [
+            self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME,
+            self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME,
+        ]);
+        $course = $this->courseProvider->getCourseById('2', $options);
+        $this->assertSame('2', $course->getIdentifier());
+        $this->assertSame('Komputationswissenschaft', $course->getName());
+        $this->assertSame('UE', $course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
+        $this->assertSame(PublicRestCourseApi::getSemesterKeys()[0],
+            $course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
+    }
+
+    public function testGetCourseByIdWithLecturersLocalDataAttribute(): void
+    {
+        $this->setUpPublicRestApi();
+
+        $courseUid = '3';
         $coResponses = [
             new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
                 json_encode([
-                    'uid' => '240759',
-                    'title' => [
-                        'value' => [
-                            'en' => 'Computational Intelligence',
-                            'de' => 'Komputationswissenschaft',
+                    'items' => [
+                        [
+                            'uid' => '42',
+                            'functionKey' => 'PRIM_LEC',
+                            'personUid' => 'DEADBEEF2',
+                            'courseUid' => $courseUid,
+                        ],
+                        [
+                            'uid' => '43',
+                            'functionKey' => 'SEC_LEC',
+                            'personUid' => 'DEADBEEF3',
+                            'courseUid' => $courseUid,
                         ],
                     ],
                 ])),
         ];
 
+        $this->mockResponses($coResponses);
+
+        $options = [];
+        Options::requestLocalDataAttributes($options, [CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE]);
+        Options::setLanguage($options, 'de');
+        $course = $this->courseProvider->getCourseById($courseUid, $options);
+
+        $this->assertSame($courseUid, $course->getIdentifier());
+        $this->assertSame('Komputationshalbwissenschaft', $course->getName());
+        $this->assertSame('45_A', $course->getCode());
+        $this->assertSame(['DEADBEEF2', 'DEADBEEF3'],
+            $course->getLocalDataValue(CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE));
+    }
+
+    public function testGetCoursesDe(): void
+    {
+        $this->setUpPublicRestApi();
+
         $options = [];
         Options::setLanguage($options, 'de');
-        $this->mockResponses($coResponses, true);
-        $course = $this->courseProvider->getCourseById('240759', $options);
-        $this->assertSame('240759', $course->getIdentifier());
+        $courses = $this->courseProvider->getCourses(1, 30, $options);
+        $this->assertCount(3, $courses);
+        $course = $courses[0];
+        $this->assertSame('1', $course->getIdentifier());
+        $this->assertSame('Komputationsbesserwissenschaft', $course->getName());
+        $this->assertSame('44_A', $course->getCode());
+        $course = $courses[1];
+        $this->assertSame('2', $course->getIdentifier());
         $this->assertSame('Komputationswissenschaft', $course->getName());
+        $this->assertSame('44_B', $course->getCode());
+        $course = $courses[2];
+        $this->assertSame('3', $course->getIdentifier());
+        $this->assertSame('Komputationshalbwissenschaft', $course->getName());
+        $this->assertSame('45_A', $course->getCode());
+    }
+
+    public function testGetCoursesEnWithLocalData(): void
+    {
+        $this->setUpPublicRestApi();
+
+        $options = [];
+        Options::requestLocalDataAttributes($options, [
+            self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME,
+            self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME,
+        ]);
+        Options::setLanguage($options, 'en');
+        $courses = $this->courseProvider->getCourses(1, 30, $options);
+        $this->assertCount(3, $courses);
+        $course = $courses[0];
+        $this->assertSame('1', $course->getIdentifier());
+        $this->assertSame('Computational Intelligence', $course->getName());
+        $this->assertSame('44_A', $course->getCode());
+        $this->assertSame('VO', $course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
+        $this->assertSame(PublicRestCourseApi::getSemesterKeys()[0],
+            $course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
+        $course = $courses[1];
+        $this->assertSame('2', $course->getIdentifier());
+        $this->assertSame('Computational Science', $course->getName());
+        $this->assertSame('44_B', $course->getCode());
+        $this->assertSame('UE', $course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
+        $this->assertSame(PublicRestCourseApi::getSemesterKeys()[0],
+            $course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
+        $course = $courses[2];
+        $this->assertSame('3', $course->getIdentifier());
+        $this->assertSame('Computational Unintelligence', $course->getName());
+        $this->assertSame('45_A', $course->getCode());
+        $this->assertSame('SEM', $course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
+        $this->assertSame(PublicRestCourseApi::getSemesterKeys()[1],
+            $course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
+    }
+
+    public function testGetCoursesEnWithLocalDataLecturers(): void
+    {
+        $this->setUpPublicRestApi();
+
+        $coResponses = [
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [
+                        [
+                            'uid' => '41',
+                            'functionKey' => 'PRIM_LEC',
+                            'personUid' => 'DEADBEEF1',
+                            'courseUid' => '1',
+                        ],
+                    ],
+                ])),
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [], // no lecturers for course 2
+                ])),
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [
+                        [
+                            'uid' => '42',
+                            'functionKey' => 'PRIM_LEC',
+                            'personUid' => 'DEADBEEF2',
+                            'courseUid' => '3',
+                        ],
+                        [
+                            'uid' => '43',
+                            'functionKey' => 'SEC_LEC',
+                            'personUid' => 'DEADBEEF3',
+                            'courseUid' => '3',
+                        ],
+                    ],
+                ])),
+        ];
+
+        $this->mockResponses($coResponses);
+
+        $options = [];
+        Options::requestLocalDataAttributes($options, [
+            CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE,
+        ]);
+        Options::setLanguage($options, 'en');
+        $courses = $this->courseProvider->getCourses(1, 30, $options);
+        $this->assertCount(3, $courses);
+        $course = $courses[0];
+        $this->assertSame('1', $course->getIdentifier());
+        $this->assertSame(['DEADBEEF1'], $course->getLocalDataValue(CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE));
+        $course = $courses[1];
+        $this->assertSame('2', $course->getIdentifier());
+        $this->assertSame([], $course->getLocalDataValue(CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE));
+        $course = $courses[2];
+        $this->assertSame('3', $course->getIdentifier());
+        $this->assertSame(['DEADBEEF2', 'DEADBEEF3'],
+            $course->getLocalDataValue(CourseEventSubscriber::LECTURERS_LOCAL_DATA_ATTRIBUTE));
     }
 
     public function testGetCoursesLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -169,7 +293,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCoursesPaginationLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -206,7 +330,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourses500Legacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(500, ['Content-Type' => 'text/xml;charset=utf-8'], ''),
@@ -218,7 +342,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCoursesInvalidXMLLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -235,7 +359,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseByIdLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -251,7 +375,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseByIdWithLecturersLocalDataAttributeLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -271,7 +395,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseByIdNotFoundLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -289,7 +413,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseById503Legacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(503, ['Content-Type' => 'text/xml;charset=utf-8'], ''),
@@ -306,7 +430,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCoursesByOrganizationLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -324,7 +448,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCourseByIdLocalDataLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -333,19 +457,19 @@ class CourseProviderTest extends ApiTestCase
 
         $options = [];
         Options::requestLocalDataAttributes($options,
-            [self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME, self::TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME]);
+            [self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME, self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME]);
         $course = $this->courseProvider->getCourseById('240759', $options);
 
         $this->assertSame('240759', $course->getIdentifier());
         $this->assertSame('Computational Intelligence', $course->getName());
         $this->assertSame('UE', $course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
         $this->assertSame('Sommersemester 2021',
-            $course->getLocalDataValue(self::TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME));
+            $course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
     }
 
     public function testGetCoursesLocalDataLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -354,19 +478,19 @@ class CourseProviderTest extends ApiTestCase
 
         $options = [];
         Options::requestLocalDataAttributes($options,
-            [self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME, self::TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME]);
+            [self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME, self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME]);
         $courses = $this->courseProvider->getCourses(1, 50, $options);
         $this->assertCount(34, $courses);
 
         foreach ($courses as $course) {
             $this->assertNotNull($course->getLocalDataValue(self::COURSE_TYPE_LOCAL_DATA_ATTRIBUTE_NAME));
-            $this->assertNotNull($course->getLocalDataValue(self::TEACHING_TERM_LOCAL_DATA_ATTRIBUTE_NAME));
+            $this->assertNotNull($course->getLocalDataValue(self::SEMESTER_LOCAL_DATA_ATTRIBUTE_NAME));
         }
     }
 
     public function testGetCoursesSearchParameterLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -417,7 +541,7 @@ class CourseProviderTest extends ApiTestCase
 
     public function testGetCoursesSearchParameterPaginationLegacy(): void
     {
-        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->setUpLegacyApi();
 
         $this->mockResponses([
             new Response(200, ['Content-Type' => 'text/xml;charset=utf-8'],
@@ -459,5 +583,139 @@ class CourseProviderTest extends ApiTestCase
             new Response(200, ['Content-Type' => 'application/json;charset=utf-8'], '{"token_endpoint": "https://token-endpoint.net/"}'),
             new Response(200, ['Content-Type' => 'application/json;charset=utf-8'], '{"access_token": "token", "expires_in": 3600, "token_type": "Bearer"}'),
         ];
+    }
+
+    private function recreateCourseCache(): void
+    {
+        $semesterKeys = PublicRestCourseApi::getSemesterKeys();
+        $coResponses = [
+            // first semester responses:
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [
+                        [
+                            'uid' => '1',
+                            'courseCode' => '44_A',
+                            'courseTypeKey' => 'VO',
+                            'semesterKey' => $semesterKeys[0],
+                            'title' => [
+                                'value' => [
+                                    'en' => 'Computational Intelligence',
+                                    'de' => 'Komputationsbesserwissenschaft',
+                                ],
+                            ],
+                        ],
+                    ],
+                    'nextCursor' => 'cursor123',
+                ])),
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [
+                        [
+                            'uid' => '2',
+                            'courseCode' => '44_B',
+                            'courseTypeKey' => 'UE',
+                            'semesterKey' => $semesterKeys[0],
+                            'title' => [
+                                'value' => [
+                                    'en' => 'Computational Science',
+                                    'de' => 'Komputationswissenschaft',
+                                ],
+                            ],
+                        ],
+                    ],
+                ])),
+            // second semester response:
+            new Response(200, ['Content-Type' => 'applicateion/json;charset=utf-8'],
+                json_encode([
+                    'items' => [
+                        [
+                            'uid' => '3',
+                            'courseCode' => '45_A',
+                            'courseTypeKey' => 'SEM',
+                            'semesterKey' => $semesterKeys[1],
+                            'title' => [
+                                'value' => [
+                                    'en' => 'Computational Unintelligence',
+                                    'de' => 'Komputationshalbwissenschaft',
+                                ],
+                            ],
+                        ],
+                    ],
+                ])),
+        ];
+        $this->mockResponses($coResponses, true);
+        try {
+            $this->courseProvider->recreateCoursesCache();
+        } catch (\Throwable $exception) { // this is expected to not fail, since sqlite does not support some operations
+            $coursesLiveTable = CachedCourse::TABLE_NAME;
+            $coursesStagingTable = CachedCourse::STAGING_TABLE_NAME;
+            $coursesTempTable = 'courses_old';
+            $courseTitlesLiveTable = CachedCourseTitle::TABLE_NAME;
+            $courseTitlesStagingTable = CachedCourseTitle::STAGING_TABLE_NAME;
+            $courseTitlesTempTable = 'course_titles_old';
+            $connection = $this->entityManager->getConnection();
+            $connection->executeStatement("ALTER TABLE $coursesLiveTable RENAME TO $coursesTempTable;");
+            $connection->executeStatement("ALTER TABLE $coursesStagingTable RENAME TO $coursesLiveTable;");
+            $connection->executeStatement("ALTER TABLE $coursesTempTable RENAME TO $coursesStagingTable;");
+            $connection->executeStatement("ALTER TABLE $courseTitlesLiveTable RENAME TO $courseTitlesTempTable;");
+            $connection->executeStatement("ALTER TABLE $courseTitlesStagingTable RENAME TO $courseTitlesLiveTable;");
+            $connection->executeStatement("ALTER TABLE $courseTitlesTempTable RENAME TO $courseTitlesStagingTable;");
+        }
+    }
+
+    private function createStagingTables(): void
+    {
+        $coursesTableName = CachedCourse::TABLE_NAME;
+        $coursesStagingTableName = CachedCourse::STAGING_TABLE_NAME;
+
+        $this->entityManager->getConnection()->executeStatement(
+            "CREATE TABLE $coursesStagingTableName AS SELECT * FROM $coursesTableName");
+
+        $courseTitlesTableName = CachedCourseTitle::TABLE_NAME;
+        $courseTitlesStagingTableName = CachedCourseTitle::STAGING_TABLE_NAME;
+
+        $this->entityManager->getConnection()->executeStatement(
+            "CREATE TABLE $courseTitlesStagingTableName AS SELECT * FROM $courseTitlesTableName");
+    }
+
+    private function setUpLegacyApi(): void
+    {
+        $this->courseProvider->setConfig($this->getLegacyApiConfig());
+        $this->courseProvider->reset();
+        $this->courseEventSubscriber->setConfig(self::createEventSubscriberConfig(false));
+    }
+
+    private function setUpPublicRestApi(): void
+    {
+        $this->courseProvider->setConfig($this->getPublicRestApiConfig());
+        $this->courseProvider->reset();
+        $this->courseEventSubscriber->setConfig(self::createEventSubscriberConfig(true));
+
+        $this->recreateCourseCache();
+    }
+
+    private function getLegacyApiConfig(): array
+    {
+        $config = [];
+        $config[Configuration::CAMPUS_ONLINE_NODE] = [
+            'legacy' => true,
+            'org_root_id' => '1',
+        ];
+
+        return $config;
+    }
+
+    private function getPublicRestApiConfig(): array
+    {
+        $config = [];
+        $config[Configuration::CAMPUS_ONLINE_NODE] = [
+            'legacy' => false,
+            'base_url' => 'https://campusonline.at/campusonline/ws/public/rest/',
+            'client_id' => 'client',
+            'client_secret' => 'secret',
+        ];
+
+        return $config;
     }
 }